# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Example Swagger setup for Golang Gin.",
        "title": "Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9876",
    "basePath": "/",
    "paths": {
...
        "/sys_server/count": {
            "get": {
                "description": "获取系统中所有服务器记录数量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "统计服务器数量",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_server/get/{id}": {
            "get": {
                "description": "根据服务器ID获取服务器信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "获取服务器详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_server/list": {
            "get": {
                "description": "支持分页、排序的服务器列表查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "获取服务器列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_server/search_by_idc": {
            "post": {
                "description": "根据IDC精确查询服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "通过IDC搜索服务器",
                "parameters": [
                    {
                        "description": "搜索参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/assistant_internal_app_repo.SearchSysServersByIDCParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_server/search_by_idc_and_svr_ip": {
            "post": {
                "description": "IDC和IP组合模糊查询服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "通过IDC和IP搜索服务器",
                "parameters": [
                    {
                        "description": "搜索参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/assistant_internal_app_repo.SearchSysServersByIDCAndSvrIPParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_server/search_by_svr_ip": {
            "post": {
                "description": "根据IP模糊查询服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器管理"
                ],
                "summary": "通过IP搜索服务器",
                "parameters": [
                    {
                        "description": "搜索参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/assistant_internal_app_repo.SearchSysServersBySvrIPParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "assistant_internal_app_repo.SearchSysServersByIDCAndSvrIPParams": {
            "type": "object",
            "properties": {
                "idc": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "svr_ip": {
                    "type": "string"
                }
            }
        },
        "assistant_internal_app_repo.SearchSysServersByIDCParams": {
            "type": "object",
            "properties": {
                "idc": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "assistant_internal_app_repo.SearchSysServersBySvrIPParams": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "svr_ip": {
                    "type": "string"
                }
            }
        },
...
        "assistant_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}