# internal/app/repo/sys_server.sql.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0
// source: sys_server.sql

package repo

import (
	"context"
	"database/sql"
)

const countSysServers = `-- name: CountSysServers :one
SELECT count(*) ct FROM sys_server
`

func (q *Queries) CountSysServers(ctx context.Context) (int64, error) {
	row := q.db.QueryRowContext(ctx, countSysServers)
	var ct int64
	err := row.Scan(&ct)
	return ct, err
}

const createSysServer = `-- name: CreateSysServer :execresult
INSERT IGNORE INTO sys_server (idc, svr_ip, svr_status, cpu_usage, mem_usage)
VALUES ('', ?, ?, 0, 0)
`

type CreateSysServerParams struct {
	SvrIP     string `json:"svr_ip"`
	SvrStatus string `json:"svr_status"`
}

func (q *Queries) CreateSysServer(ctx context.Context, arg CreateSysServerParams) (sql.Result, error) {
	return q.db.ExecContext(ctx, createSysServer, arg.SvrIP, arg.SvrStatus)
}

const deleteSysServerByID = `-- name: DeleteSysServerByID :execresult
DELETE FROM sys_server
WHERE id = ?
LIMIT 1
`

func (q *Queries) DeleteSysServerByID(ctx context.Context, id int64) (sql.Result, error) {
	return q.db.ExecContext(ctx, deleteSysServerByID, id)
}

const deleteSysServerBySvrIP = `-- name: DeleteSysServerBySvrIP :execresult
DELETE FROM sys_server
WHERE svr_ip = ?
LIMIT 1
`

func (q *Queries) DeleteSysServerBySvrIP(ctx context.Context, svrIp string) (sql.Result, error) {
	return q.db.ExecContext(ctx, deleteSysServerBySvrIP, svrIp)
}

const getSysServerByID = `-- name: GetSysServerByID :one
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
WHERE id = ?
LIMIT 1
`

func (q *Queries) GetSysServerByID(ctx context.Context, id int64) (SysServer, error) {
	row := q.db.QueryRowContext(ctx, getSysServerByID, id)
	var i SysServer
	err := row.Scan(
		&i.ID,
		&i.GmtCreate,
		&i.GmtModified,
		&i.IDC,
		&i.SvrIP,
		&i.SvrStatus,
		&i.CpuUsage,
		&i.MemUsage,
	)
	return i, err
}

const getSysServerBySvrIP = `-- name: GetSysServerBySvrIP :one
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
WHERE svr_ip = ?
LIMIT 1
`

func (q *Queries) GetSysServerBySvrIP(ctx context.Context, svrIp string) (SysServer, error) {
	row := q.db.QueryRowContext(ctx, getSysServerBySvrIP, svrIp)
	var i SysServer
	err := row.Scan(
		&i.ID,
		&i.GmtCreate,
		&i.GmtModified,
		&i.IDC,
		&i.SvrIP,
		&i.SvrStatus,
		&i.CpuUsage,
		&i.MemUsage,
	)
	return i, err
}

const listSysServers = `-- name: ListSysServers :many
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
ORDER BY id DESC
LIMIT ? OFFSET ?
`

type ListSysServersParams struct {
	Limit  int32 `json:"limit"`
	Offset int32 `json:"offset"`
}

func (q *Queries) ListSysServers(ctx context.Context, arg ListSysServersParams) ([]SysServer, error) {
	rows, err := q.db.QueryContext(ctx, listSysServers, arg.Limit, arg.Offset)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	items := []SysServer{}
	for rows.Next() {
		var i SysServer
		if err := rows.Scan(
			&i.ID,
			&i.GmtCreate,
			&i.GmtModified,
			&i.IDC,
			&i.SvrIP,
			&i.SvrStatus,
			&i.CpuUsage,
			&i.MemUsage,
		); err != nil {
			return nil, err
		}
		items = append(items, i)
	}
	if err := rows.Close(); err != nil {
		return nil, err
	}
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return items, nil
}

const searchSysServersByIDC = `-- name: SearchSysServersByIDC :many
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
WHERE idc = ?
ORDER BY id DESC
LIMIT ? OFFSET ?
`

type SearchSysServersByIDCParams struct {
	IDC    string `json:"idc"`
	Limit  int32  `json:"limit"`
	Offset int32  `json:"offset"`
}

func (q *Queries) SearchSysServersByIDC(ctx context.Context, arg SearchSysServersByIDCParams) ([]SysServer, error) {
	rows, err := q.db.QueryContext(ctx, searchSysServersByIDC, arg.IDC, arg.Limit, arg.Offset)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	items := []SysServer{}
	for rows.Next() {
		var i SysServer
		if err := rows.Scan(
			&i.ID,
			&i.GmtCreate,
			&i.GmtModified,
			&i.IDC,
			&i.SvrIP,
			&i.SvrStatus,
			&i.CpuUsage,
			&i.MemUsage,
		); err != nil {
			return nil, err
		}
		items = append(items, i)
	}
	if err := rows.Close(); err != nil {
		return nil, err
	}
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return items, nil
}

const searchSysServersByIDCAndSvrIP = `-- name: SearchSysServersByIDCAndSvrIP :many
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
WHERE idc like ?
  AND svr_ip like ?
ORDER BY id DESC
LIMIT ? OFFSET ?
`

type SearchSysServersByIDCAndSvrIPParams struct {
	IDC    string `json:"idc"`
	SvrIP  string `json:"svr_ip"`
	Limit  int32  `json:"limit"`
	Offset int32  `json:"offset"`
}

func (q *Queries) SearchSysServersByIDCAndSvrIP(ctx context.Context, arg SearchSysServersByIDCAndSvrIPParams) ([]SysServer, error) {
	rows, err := q.db.QueryContext(ctx, searchSysServersByIDCAndSvrIP,
		arg.IDC,
		arg.SvrIP,
		arg.Limit,
		arg.Offset,
	)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	items := []SysServer{}
	for rows.Next() {
		var i SysServer
		if err := rows.Scan(
			&i.ID,
			&i.GmtCreate,
			&i.GmtModified,
			&i.IDC,
			&i.SvrIP,
			&i.SvrStatus,
			&i.CpuUsage,
			&i.MemUsage,
		); err != nil {
			return nil, err
		}
		items = append(items, i)
	}
	if err := rows.Close(); err != nil {
		return nil, err
	}
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return items, nil
}

const searchSysServersBySvrIP = `-- name: SearchSysServersBySvrIP :many
SELECT
  id,
  gmt_create,
  gmt_modified,
  idc,
  svr_ip,
  svr_status,
  cpu_usage,
  mem_usage
FROM sys_server
WHERE svr_ip like ?
ORDER BY id DESC
LIMIT ? OFFSET ?
`

type SearchSysServersBySvrIPParams struct {
	SvrIP  string `json:"svr_ip"`
	Limit  int32  `json:"limit"`
	Offset int32  `json:"offset"`
}

func (q *Queries) SearchSysServersBySvrIP(ctx context.Context, arg SearchSysServersBySvrIPParams) ([]SysServer, error) {
	rows, err := q.db.QueryContext(ctx, searchSysServersBySvrIP, arg.SvrIP, arg.Limit, arg.Offset)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	items := []SysServer{}
	for rows.Next() {
		var i SysServer
		if err := rows.Scan(
			&i.ID,
			&i.GmtCreate,
			&i.GmtModified,
			&i.IDC,
			&i.SvrIP,
			&i.SvrStatus,
			&i.CpuUsage,
			&i.MemUsage,
		); err != nil {
			return nil, err
		}
		items = append(items, i)
	}
	if err := rows.Close(); err != nil {
		return nil, err
	}
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return items, nil
}

const updateSysServerMetricsBySvrIP = `-- name: UpdateSysServerMetricsBySvrIP :execresult
UPDATE sys_server
SET cpu_usage = ?,
    mem_usage = ?,
    svr_status = ?
WHERE svr_ip = ?
LIMIT 1
`

type UpdateSysServerMetricsBySvrIPParams struct {
	CpuUsage  float64 `json:"cpu_usage"`
	MemUsage  float64 `json:"mem_usage"`
	SvrStatus string  `json:"svr_status"`
	SvrIP     string  `json:"svr_ip"`
}

func (q *Queries) UpdateSysServerMetricsBySvrIP(ctx context.Context, arg UpdateSysServerMetricsBySvrIPParams) (sql.Result, error) {
	return q.db.ExecContext(ctx, updateSysServerMetricsBySvrIP,
		arg.CpuUsage,
		arg.MemUsage,
		arg.SvrStatus,
		arg.SvrIP,
	)
}